added endpoints, new repository's methods

+++ src/main/java/net/thumbtack/onlineshop/entity/Administrator.java
@@ -7,7 +7,7 @@ public class Administrator extends User {
     public Administrator() {
     }
 
-    public Administrator(long id, String firstName, String lastName, String patronymic, String login,
+    public Administrator(int id, String firstName, String lastName, String patronymic, String login,
                          String password, String position) {
         super(id, firstName, lastName, patronymic, login, password);
         this.position = position;
